# services/api-gateway/api/openapi.yaml
openapi: 3.0.3
info:
  title: API Gateway
  version: 1.0.0
  description: |
    Единая точка входа в интернет-магазин. Все клиентские запросы идут через
    API Gateway, который проверяет JWT-токен, подставляет заголовки
    `X-User-Id` / `X-User-Roles` и проксирует вызовы в микросервисы:

    | Раздел API        | Микросервис     | Транспорт |
    |-------------------|-----------------|-----------|
    | `/auth/*`         | auth-service    | HTTP      |
    | `/users/*`        | user-service    | HTTP      |
    | `/products/*`     | product-service | HTTP      |
    | `/categories/*`   | product-service | HTTP      |
    | `/inventory/*`    | inventory-service | HTTP    |
    | `/orders/*`       | order-service   | HTTP      |
    | остатки в каталоге | inventory-service | gRPC    |

    ## Как тестировать

    1. `POST /api/v1/auth/register` — регистрация (получите письмо с токеном подтверждения).
    2. `POST /api/v1/auth/login` — вход. В ответе придёт `access_token` и cookie `refresh_token`.
    3. Нажмите **Authorize** в Swagger UI и вставьте `access_token` (без слова `Bearer`).
    4. Все защищённые эндпоинты станут доступны.

    ## Авторизация

    - Защищённые эндпоинты требуют заголовок `Authorization: Bearer <access_token>`.
    - `access_token` — короткоживущий JWT (RS256, issuer `auth-service`).
    - `refresh_token` передаётся в HttpOnly-cookie `refresh_token`
      и используется только в `POST /api/v1/auth/refresh` и `POST /api/v1/auth/logout`.
...

...

...

...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/orders:
    post:
      summary: Создать заказ
      description: |
        Оформляет заказ от имени пользователя из access-токена.

        **Обязательный заголовок `Idempotency-Key`** — UUID. Повторный запрос
        с тем же ключом возвращает результат первого запроса, а не создаёт
        новый заказ. Для каждой новой попытки оформления нужен новый ключ
        (например, сгенерируйте его кнопкой в Swagger UI или
        `uuidgen`).

        Операция выполняется через Saga: резервирование остатков,
        списание, подтверждение оплаты. Если что-то не удалось —
        выполняется компенсация, а заказ получает статус `CANCELLED`.
        Итоговый статус возвращается в поле `status`.
      tags: [orders]
      operationId: createOrder
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
            examples:
              default:
                value:
                  items:
                    - product_id: 507f1f77bcf86cd799439011
                      quantity: 2
                    - product_id: 507f1f77bcf86cd799439012
                      quantity: 1
      responses:
        '201':
          description: Заказ создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateOrderResponse'
              example:
                order_id: 6ba7b810-9dad-11d1-80b4-00c04fd430c8
                status: RESERVED
                total_price: 21999800
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Товар из списка не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Недостаточно остатков или заказ с таким `Idempotency-Key` уже обработан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      summary: Список своих заказов
      description: |
        Возвращает заказы текущего пользователя, отсортированные
        от новых к старым. Ответ — массив объектов, пагинации нет.
      tags: [orders]
      operationId: listOrders
      responses:
        '200':
          description: Список заказов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
              example:
                - order_id: 6ba7b810-9dad-11d1-80b4-00c04fd430c8
                  user_id: 550e8400-e29b-41d4-a716-446655440000
                  status: PAID
                  total_price: 21999800
                  created_at: '2024-02-01T12:00:00Z'
                  updated_at: '2024-02-01T12:05:00Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/orders/{id}/ship:
    parameters:
      - $ref: '#/components/parameters/OrderIdPath'

    post:
      summary: Перевести заказ в статус «Отправлен»
      description: |
        Переход статуса `PAID` → `SHIPPED`.
        Требуется `ROLE_MANAGER` или `ROLE_ADMIN`.
        Недопустимый переход возвращает `409`.
      tags: [orders]
      operationId: shipOrder
      responses:
        '204':
          description: Заказ отправлен
        '400':
          description: "`id` не является корректным UUID"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Заказ не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Недопустимый переход статуса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/orders/{id}/complete:
    parameters:
      - $ref: '#/components/parameters/OrderIdPath'

    post:
      summary: Завершить заказ
      description: |
        Переход статуса `SHIPPED` → `COMPLETED`.
        Требуется `ROLE_MANAGER` или `ROLE_ADMIN`.
      tags: [orders]
      operationId: completeOrder
      responses:
        '204':
          description: Заказ завершён
        '400':
          description: "`id` не является корректным UUID"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Заказ не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Недопустимый переход статуса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/orders/{id}/payment/success:
    parameters:
      - $ref: '#/components/parameters/OrderIdPath'

    post:
      summary: Подтверждение успешной оплаты
      description: |
        Обрабатывает событие об успешной оплате: переводит заказ
        в статус `PAID`. Идемпотентна по `event_id` — повторная доставка
        того же события не меняет состояние заказа повторно.

        Требуется `ROLE_MANAGER` или `ROLE_ADMIN`.
      tags: [orders]
      operationId: paymentSuccess
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentSuccessRequest'
            examples:
              default:
                value:
                  event_id: 7c9e6679-7425-40de-944b-e07fc1f90ae7
                  correlation_id: 1b671a64-40d5-491e-99b0-da01ff1f3341
      responses:
        '204':
          description: Оплата подтверждена
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Заказ не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Недопустимый переход статуса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/orders/{id}/payment/failed:
    parameters:
      - $ref: '#/components/parameters/OrderIdPath'

    post:
      summary: Сообщение об отказе оплаты
      description: |
        Обрабатывает событие об отказе оплаты: отменяет заказ и
        возвращает зарезервированные остатки на склад.
        Идемпотентна по `event_id`.

        Требуется `ROLE_MANAGER` или `ROLE_ADMIN`.
      tags: [orders]
      operationId: paymentFailed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentFailedRequest'
            examples:
              default:
                value:
                  event_id: 7c9e6679-7425-40de-944b-e07fc1f90ae7
                  correlation_id: 1b671a64-40d5-491e-99b0-da01ff1f3341
                  reason: insufficient funds
      responses:
        '204':
          description: Отказ по оплате обработан, заказ отменён
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Заказ не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Недопустимый переход статуса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Access-токен, полученный в `POST /api/v1/auth/login`
        или `POST /api/v1/auth/refresh`.

        В Swagger UI нажмите **Authorize** и вставьте только значение
        токена, без префикса `Bearer `.

        Получить токен:
        ```bash
        curl -s -X POST http://localhost:4000/api/v1/auth/login \
          -H 'Content-Type: application/json' \
          -d '{"email":"user@example.com","password":"password123"}'
        ```

        Токен подписан RS256, issuer — `auth-service`. Gateway проверяет
        подпись, срок действия и `exp`/`nbf` с небольшим допуском (leeway).
        Истёкший или подделанный токен → `401`.
    refreshTokenCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: |
        HttpOnly-cookie с refresh-токеном. Выставляется автоматически
        ответами `login`, `oauth/callback` и `refresh`.

        Cookie имеет флаги `Secure` и `Path=/api/v1/auth/refresh`,
        поэтому по `http://localhost:4000` браузер её не сохранит —
        для локального тестирования задайте значение cookie вручную
        через **Authorize** в Swagger UI.
  parameters:
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Ключ идемпотентности (UUID) для создания заказа.
        Повторный запрос с тем же ключом не создаёт новый заказ,
        а возвращает результат первого.
      schema:
        type: string
        format: uuid
      example: 6ba7b810-9dad-11d1-80b4-00c04fd430c8
...
    OrderIdPath:
      name: id
      in: path
      required: true
      description: ID заказа (UUID)
      schema:
        type: string
        format: uuid
      example: 6ba7b810-9dad-11d1-80b4-00c04fd430c8
...
  schemas:
    ErrorResponse:
      type: object
      description: Единый формат ошибки для всех эндпоинтов шлюза
      required: [error]
      properties:
        error:
          type: string
          description: Сообщение об ошибке (может быть на русском языке)
          example: Проверьте корректность данных
...
    OrderItem:
      type: object
      required: [product_id, quantity]
      properties:
        product_id:
          type: string
          description: ID товара из каталога
          example: 507f1f77bcf86cd799439011
        quantity:
          type: integer
          minimum: 1
          example: 2
    CreateOrderRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          description: Состав заказа. Повторяющиеся товары объединяются сервисом.
          items:
            $ref: '#/components/schemas/OrderItem'
    OrderStatus:
      type: string
      description: |
        Статус заказа:
        - `NEW` — создан
        - `RESERVED` — остатки зарезервированы, ожидает оплаты
        - `PAID` — оплата подтверждена
        - `SHIPPED` — передан в доставку
        - `COMPLETED` — вручён
        - `CANCELLED` — отменён (в т.ч. по ошибке Saga)
      enum: [NEW, RESERVED, PAID, SHIPPED, COMPLETED, CANCELLED]
    CreateOrderResponse:
      type: object
      required: [order_id, status, total_price]
      properties:
        order_id:
          type: string
          format: uuid
          example: 6ba7b810-9dad-11d1-80b4-00c04fd430c8
        status:
          $ref: '#/components/schemas/OrderStatus'
        total_price:
          type: integer
          format: int64
          description: Итоговая сумма заказа в копейках
          example: 21999800
    Order:
      type: object
      required: [order_id, user_id, status, total_price, created_at, updated_at]
      properties:
        order_id:
          type: string
          format: uuid
          example: 6ba7b810-9dad-11d1-80b4-00c04fd430c8
        user_id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        status:
          $ref: '#/components/schemas/OrderStatus'
        total_price:
          type: integer
          format: int64
          description: Итоговая сумма заказа в копейках
          example: 21999800
        created_at:
          type: string
          format: date-time
          example: '2024-02-01T12:00:00Z'
        updated_at:
          type: string
          format: date-time
          example: '2024-02-01T12:05:00Z'
    PaymentSuccessRequest:
      type: object
      required: [event_id, correlation_id]
      properties:
        event_id:
          type: string
          format: uuid
          description: ID события оплаты — ключ идемпотентности обработки
          example: 7c9e6679-7425-40de-944b-e07fc1f90ae7
        correlation_id:
          type: string
          format: uuid
          description: ID корреляции для прослеживания по логам и трейсам
          example: 1b671a64-40d5-491e-99b0-da01ff1f3341
    PaymentFailedRequest:
      type: object
      required: [event_id, correlation_id, reason]
      properties:
        event_id:
          type: string
          format: uuid
          description: ID события оплаты — ключ идемпотентности обработки
          example: 7c9e6679-7425-40de-944b-e07fc1f90ae7
        correlation_id:
          type: string
          format: uuid
          description: ID корреляции для прослеживания по логам и трейсам
          example: 1b671a64-40d5-491e-99b0-da01ff1f3341
        reason:
          type: string
          description: Причина отказа (например, недостаточно средств)
          example: insufficient funds
  responses:
    BadRequest:
      description: Некорректный запрос (невалидный UUID, отсутствие обязательного параметра)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: invalid id format
    ValidationError:
      description: Ошибка валидации тела запроса (`Проверьте корректность данных`)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Проверьте корректность данных
    Unauthorized:
      description: |
        Не передан `Authorization: Bearer <token>`, токен невалиден или истёк
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            noToken:
              summary: Токен не передан
              value:
                error: missing access token
            invalid:
              summary: Токен невалиден или истёк
              value:
                error: invalid or expired access token
    Forbidden:
      description: Недостаточно прав — у пользователя нет требуемой роли
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: insufficient permissions
...
    TooManyRequests:
      description: |
        Превышен лимит частоты запросов. Повторите через
        `X-RateLimit-Retry-After` секунд.
      headers:
        X-RateLimit-Limit:
          description: Максимальное число запросов в окне
          schema:
            type: integer
          example: 60
        X-RateLimit-Remaining:
          description: Сколько запросов осталось
          schema:
            type: integer
          example: 0
        X-RateLimit-RetryAfter:
          description: Через сколько секунд можно повторить
          schema:
            type: integer
          example: 2
        X-RateLimit-Reset:
          description: Unix-время сброса окна
          schema:
            type: integer
            format: int64
          example: 1736940000
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: too many requests
    InternalError:
      description: Внутренняя ошибка шлюза или недоступность микросервиса
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Internal server error